+++ src/events.js
@@ -1,5 +1,3 @@
-import xdr from './xdr';
-
 import { StrKey } from './strkey';
 import { scValToNative } from './scval';
 
@@ -19,7 +17,9 @@ import { scValToNative } from './scval';
  */
 export function humanizeEvents(events) {
   return events.map((e) => {
-    if (e instanceof xdr.DiagnosticEvent) {
+    // A pseudo-instanceof check for xdr.DiagnosticEvent more reliable
+    // in mixed SDK environments:
+    if (e.inSuccessfulContractCall) {
       return extractEvent(e.event());
     }
 
@@ -29,9 +29,10 @@ export function humanizeEvents(events) {
 
 function extractEvent(event) {
   return {
-    ...(event.contractId() != null && {
-      contractId: StrKey.encodeContract(event.contractId())
-    }),
+    ...(typeof event.contractId === 'function' &&
+      event.contractId() != null && {
+        contractId: StrKey.encodeContract(event.contractId())
+      }),
     type: event.type().name,
     topics: event
       .body()